feat: fix alert conditions thresholds
We only need the signal loss (aka. expiration), since for the queries
that are used we only get values of 1 or None (aka. no signal). The
window for loss signal (i.e. how long to wait for signal to be missing
before it's considered a signal loss) should be at least 6 minutes,
since the synthetic monitors are pinging the URL once every 5 minutes
(the window should be bigger than the frequency of the pings).