fix: duplicate alert conditions are being created
In one of the previous commits we changed the template for the names of
the alert conditions. The code that checks if conditions have to be
created looks them up based on a template, that was not modified, so new
alert conditions were not being returned, and the check assumed that
alert conditions haven't been created, so duplicates were being created.